refactoring, Logger updated - log file introduced, tests refactoring
