+++ log4j-application.log
@@ -1,0 +1,26 @@
+2019-04-12 13:04:28,503  INFO Main: 
+Program has stared
+2019-04-12 13:04:32,401  WARN Main: 
+Reading sudoku - incorrect input - using log4j
+2019-04-12 13:04:37,186  INFO Main: 
+Program has finished
+2019-04-12 13:14:15,613  INFO Main: 
+Program has stared
+2019-04-12 13:14:39,645  WARN Main: 
+Reading sudoku - incorrect input - using log4j
+2019-04-12 13:14:51,709  WARN Main: 
+Reading sudoku - incorrect input - using log4j
+2019-04-12 13:15:30,974  INFO Main: 
+Program has finished
+%highlight{2019-04-12 13:21:24,909  INFO Main: 
+Program has stared
+}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}%highlight{2019-04-12 13:22:24,587  INFO Main: 
+Program has finished
+}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}%highlight{2019-04-12 13:22:27,725 INFO  [main] (Main.java:41) - Program has stared
+}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}%highlight{2019-04-12 13:22:48,538 INFO  [main] (Main.java:77) - Program has finished
+}{FATAL=red, ERROR=red, WARN=yellow bold, INFO=black, DEBUG=green bold, TRACE=blue}2019-04-12 13:22:51,761  INFO Main: 
+Program has stared
+2019-04-12 13:23:14,603  WARN Main: 
+Reading sudoku - incorrect input - using log4j
+2019-04-12 13:23:17,034  INFO Main: 
+Program has finished
