+++ src/main/java/sudoku/processing/SudokuService.java
@@ -35,45 +35,6 @@ public class SudokuService {
         this.squares = squares;
     }
 
-    public ArrayList<List<? extends SudokuElement>> createSudokuElementObjectsService(int[][] data) {
-        for (int i = 0; i < data.length; i++) {
-            Horizontal horizontal = new Horizontal();
-            for (int j = 0; j < data[i].length; j++) {
-                Cell cell = new Cell(data[i][j], i, j);
-                horizontal.getCells().add(cell);
-
-                if (i == 0) {
-                    Vertical vertical = new Vertical();
-                    verticals.add(vertical);
-                    verticals.get(j).getCells().add(cell);
-                } else {
-                    verticals.get(j).getCells().add(cell);
-                }
-
-
-                if (shouldCreateSquare(i, j)) {
-                    Square square = new Square();
-                    squares.add(square);
-                }
-
-                squares.get((i/3)*3 + j/3).getCells().add(cell);
-
-            }
-            horizontals.add(horizontal);
-        }
-        // <? extends sudoku.model.SudokuElement> pre vsetky objekty, ktore dedia od sudoku.model.SudokuElement
-        ArrayList<List<? extends SudokuElement>> sudokuElements = new ArrayList<>();
-        sudokuElements.add(horizontals);
-        sudokuElements.add(verticals);
-        sudokuElements.add(squares);
-        return sudokuElements;
-    }
-
-    // iba kvoli testu public, inak bolo private
-    public boolean shouldCreateSquare(int i, int j) {
-        return (i % 3 == 0) && (j % 3 == 0);
-    }
-
     public void resolveSudokuService() {
         System.out.println("POSSIBILITIES: ");
         Solution solution = new Solution(verticals, horizontals, squares);
