+++ src/main/resources/log4j.properties
@@ -1,0 +1,26 @@
+# Add a new category/logger called "ExternalAppLogger" and specify
+# the log level ("WARN") and the appender name ("extAppLogFile").
+# VSETKO VYSSIE AKO INFO IDE DO FILE
+log4j.category.ExternalAppLogger=INFO, extAppLogFile
+
+# Set the appender settings
+log4j.appender.extAppLogFile=org.apache.log4j.RollingFileAppender
+log4j.appender.extAppLogFile.File=C:\\Users\\lukr\\IdeaProjects\\sudoku-with-tests\\log4j-application.log
+log4j.appender.extAppLogFile.MaxFileSize=20000KB
+log4j.appender.extAppLogFile.MaxBackupIndex=10
+log4j.appender.extAppLogFile.layout=org.apache.log4j.PatternLayout
+log4j.appender.extAppLogFile.layout.ConversionPattern=%d %5p %C{1}: %n%m%n
+
+# Priority: Trace < Debug < Info < Warn < Error < Fatal
+# Root logger option
+log4j.rootLogger=WARN, myInfo
+
+# Redirect only WARN log messages to console (filter restricts the level accepted)
+log4j.appender.myInfo=org.apache.log4j.ConsoleAppender
+log4j.appender.myInfo.Target=System.out
+log4j.appender.myInfo.layout=org.apache.log4j.PatternLayout
+log4j.appender.myInfo.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
+log4j.appender.myInfo.filter.a=org.apache.log4j.varia.LevelRangeFilter
+log4j.appender.myInfo.filter.a.levelMin=WARN
+log4j.appender.myInfo.filter.a.levelMax=WARN
+log4j.appender.myInfo.filter.a.acceptOnMatch=true
